Backend/src/rag/requirements.txt
requests>=2.28.0
pandas>=1.4.0
fpdf>=1.7.2
chromadb>=0.4.0
sentence-transformers>=2.2.0
PyMuPDF>=1.22.0
openai>=1.0.0
python-dotenv>=0.19.0
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.27.0
anthropic>=0.18.0
tenacity>=8.2.0
